Add integration tests for partials WIP
The new tests only pass when run in isolation. When run together with
the full test suite they fail apparently because the
`CardDeck::View.configuration#root` is reset to it's default value when
the configuration is unloaded but not set back to it's 'correct' value
when reloaded. Same applies to `CardDeck::View.configuration#root`.